Fix unit tests after response refactor
