+++ matchbook-sdk-core/src/main/java/com/matchbook/sdk/core/clients/rest/UserRestClient.java
@@ -1,14 +1,14 @@
 package com.matchbook.sdk.core.clients.rest;
 
 import com.matchbook.sdk.core.StreamObserver;
-import com.matchbook.sdk.core.clients.rest.dtos.user.Login;
 import com.matchbook.sdk.core.clients.rest.dtos.user.LoginRequest;
-import com.matchbook.sdk.core.clients.rest.dtos.user.Logout;
+import com.matchbook.sdk.core.clients.rest.dtos.user.LoginResponse;
+import com.matchbook.sdk.core.clients.rest.dtos.user.LogoutResponse;
 
 public interface UserRestClient extends RestClient {
 
-    void login(LoginRequest loginRequest, StreamObserver<Login> response);
+    void login(LoginRequest loginRequest, StreamObserver<LoginResponse> response);
 
-    void logout(String sessionToken, StreamObserver<Logout> response);
+    void logout(String sessionToken, StreamObserver<LogoutResponse> response);
 
 }
