+++ matchbook-sdk-core/src/main/java/com/matchbook/sdk/core/clients/rest/UserRestClientImpl.java
@@ -4,10 +4,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.matchbook.sdk.core.StreamObserver;
-import com.matchbook.sdk.core.clients.rest.dtos.user.Login;
 import com.matchbook.sdk.core.clients.rest.dtos.user.LoginRequest;
 import com.matchbook.sdk.core.clients.rest.dtos.user.LoginResponse;
-import com.matchbook.sdk.core.clients.rest.dtos.user.Logout;
+import com.matchbook.sdk.core.clients.rest.dtos.user.LogoutResponse;
 import com.matchbook.sdk.core.configs.ClientConnectionManager;
 import com.matchbook.sdk.core.exceptions.MatchbookSDKHTTPException;
 
@@ -24,7 +23,7 @@ public class UserRestClientImpl extends AbstractRestClient implements UserRestCl
     }
 
     @Override
-    public void login(LoginRequest loginRequest, StreamObserver<Login> loginObserver) {
+    public void login(LoginRequest loginRequest, StreamObserver<LoginResponse> loginObserver) {
         try {
             String path = clientConnectionManager.getClientConfig().getLoginUrl();
             String requestBody = loginRequestWriter.writeValueAsString(loginRequest);
@@ -35,7 +34,7 @@ public class UserRestClientImpl extends AbstractRestClient implements UserRestCl
     }
 
     @Override
-    public void logout(String sessionToken, StreamObserver<Logout> response) {
+    public void logout(String sessionToken, StreamObserver<LogoutResponse> response) {
         throw new UnsupportedOperationException("Not implemented");
     }
