--- matchbook-sdk-core/src/main/java/com/matchbook/sdk/core/clients/rest/dtos/user/Login.java
@@ -1,30 +1,0 @@
-package com.matchbook.sdk.core.clients.rest.dtos.user;
-
-public class Login {
-
-    private String sessionToken;
-    private Long userId;
-    private Account account;
-
-    public String getSessionToken() {
-        return sessionToken;
-    }
-
-    public long getUserId() {
-        return userId;
-    }
-
-    public Account getAccount() {
-        return account;
-    }
-
-    @Override
-    public String toString() {
-        return Login.class.getSimpleName() + " {" +
-                "sessionToken=" + sessionToken +
-                ", userId=" + userId +
-                ", account=" + account +
-                "}";
-    }
-
-}
