--- matchbook-sdk-core/src/main/java/com/matchbook/sdk/core/clients/rest/dtos/user/Logout.java
@@ -1,29 +1,0 @@
-package com.matchbook.sdk.core.clients.rest.dtos.user;
-
-public class Logout {
-
-    private String sessionToken;
-    private Long userId;
-    private String username;
-
-    public String getSessionToken() {
-        return sessionToken;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    @Override
-    public String toString() {
-        return Logout.class.getSimpleName() + " {" +
-                "sessionToken=" + sessionToken +
-                ", userId=" + userId +
-                ", username=" + username +
-                "}";
-    }
-}
